Review and pull request with improvements
