+++ numbers.json
@@ -1,7 +1,7 @@
 [
 	{
 		"Numeros" : [
-			2 , 121 , 11 , 44 , 97 , 887 , 43, 33 , 997 , 22 , 32 , 41 , 9   
+			2 , 121 , 11 , 44, "d", 97, 8 , 887 , 43, 33 , 997 , 22 , 32 , "el jaime", 41 , 9
 		]
 
 	}
